# Conflicts:
#	app/assets/javascripts/discourse/app/components/sidebar/footer.js
#	app/assets/javascripts/discourse/app/lib/register-service-worker.js

+++ spec/rails_helper.rb
@@ -1,26 +1,28 @@
 # frozen_string_literal: true
 
-if ENV['COVERAGE']
-  require 'simplecov'
-  SimpleCov.command_name "#{SimpleCov.command_name} #{ENV['TEST_ENV_NUMBER']}" if ENV['TEST_ENV_NUMBER']
-  SimpleCov.start 'rails' do
-    add_group 'Libraries', /^\/lib\/(?!tasks).*$/
-    add_group 'Scripts', 'script'
-    add_group 'Serializers', 'app/serializers'
-    add_group 'Services', 'app/services'
-    add_group 'Tasks', 'lib/tasks'
+if ENV["COVERAGE"]
+  require "simplecov"
+  if ENV["TEST_ENV_NUMBER"]
+    SimpleCov.command_name "#{SimpleCov.command_name} #{ENV["TEST_ENV_NUMBER"]}"
+  end
+  SimpleCov.start "rails" do
+    add_group "Libraries", %r{^/lib/(?!tasks).*$}
+    add_group "Scripts", "script"
+    add_group "Serializers", "app/serializers"
+    add_group "Services", "app/services"
+    add_group "Tasks", "lib/tasks"
   end
 end
 
-require 'rubygems'
-require 'rbtrace' if RUBY_ENGINE == "ruby"
-require 'pry'
-require 'pry-byebug'
-require 'pry-rails'
-require 'fabrication'
-require 'mocha/api'
-require 'certified'
-require 'webmock/rspec'
+require "rubygems"
+require "rbtrace" if RUBY_ENGINE == "ruby"
+require "pry"
+require "pry-byebug"
+require "pry-rails"
+require "fabrication"
+require "mocha/api"
+require "certified"
+require "webmock/rspec"
 
 class RspecErrorTracker
   def self.last_exception=(ex)
@@ -39,9 +41,9 @@ class RspecErrorTracker
     begin
       @app.call(env)
 
-    # This is a little repetitive, but since WebMock::NetConnectNotAllowedError
-    # and also Mocha::ExpectationError inherit from Exception instead of StandardError
-    # they do not get captured by the rescue => e shorthand :(
+      # This is a little repetitive, but since WebMock::NetConnectNotAllowedError
+      # and also Mocha::ExpectationError inherit from Exception instead of StandardError
+      # they do not get captured by the rescue => e shorthand :(
     rescue WebMock::NetConnectNotAllowedError, Mocha::ExpectationError, StandardError => e
       RspecErrorTracker.last_exception = e
       raise e
@@ -50,16 +52,16 @@ class RspecErrorTracker
   end
 end
 
-ENV["RAILS_ENV"] ||= 'test'
+ENV["RAILS_ENV"] ||= "test"
 require File.expand_path("../../config/environment", __FILE__)
-require 'rspec/rails'
-require 'shoulda-matchers'
-require 'sidekiq/testing'
-require 'test_prof/recipes/rspec/let_it_be'
-require 'test_prof/before_all/adapters/active_record'
-require 'webdrivers'
-require 'selenium-webdriver'
-require 'capybara/rails'
+require "rspec/rails"
+require "shoulda-matchers"
+require "sidekiq/testing"
+require "test_prof/recipes/rspec/let_it_be"
+require "test_prof/before_all/adapters/active_record"
+require "webdrivers"
+require "selenium-webdriver"
+require "capybara/rails"
 
 # The shoulda-matchers gem no longer detects the test framework
 # you're using or mixes itself into that framework automatically.
@@ -75,26 +77,19 @@ end
 # in spec/support/ and its subdirectories.
 Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
 
-require Rails.root.join("spec/system/page_objects/pages/base.rb")
-require Rails.root.join("spec/system/page_objects/modals/base.rb")
+Dir[Rails.root.join("spec/system/page_objects/**/base.rb")].each { |f| require f }
 Dir[Rails.root.join("spec/system/page_objects/**/*.rb")].each { |f| require f }
 
 Dir[Rails.root.join("spec/fabricators/*.rb")].each { |f| require f }
-require_relative './helpers/redis_snapshot_helper'
+require_relative "./helpers/redis_snapshot_helper"
 
 # Require plugin helpers at plugin/[plugin]/spec/plugin_helper.rb (includes symlinked plugins).
-if ENV['LOAD_PLUGINS'] == "1"
-  Dir[Rails.root.join("plugins/*/spec/plugin_helper.rb")].each do |f|
-    require f
-  end
+if ENV["LOAD_PLUGINS"] == "1"
+  Dir[Rails.root.join("plugins/*/spec/plugin_helper.rb")].each { |f| require f }
 
-  Dir[Rails.root.join("plugins/*/spec/fabricators/**/*.rb")].each do |f|
-    require f
-  end
+  Dir[Rails.root.join("plugins/*/spec/fabricators/**/*.rb")].each { |f| require f }
 
-  Dir[Rails.root.join("plugins/*/spec/system/page_objects/**/*.rb")].each do |f|
-    require f
-  end
+  Dir[Rails.root.join("plugins/*/spec/system/page_objects/**/*.rb")].each { |f| require f }
 end
 
 # let's not run seed_fu every test
@@ -106,7 +101,7 @@ SeedFu.seed
 
 # we need this env var to ensure that we can impersonate in test
 # this enable integration_helpers sign_in helper
-ENV['DISCOURSE_DEV_ALLOW_ANON_TO_IMPERSONATE'] = '1'
+ENV["DISCOURSE_DEV_ALLOW_ANON_TO_IMPERSONATE"] = "1"
 
 module TestSetup
   # This is run before each test and before each before_all block
@@ -119,9 +114,7 @@ module TestSetup
     SiteIconManager.disable
     WordWatcher.disable_cache
 
-    SiteSetting.provider.all.each do |setting|
-      SiteSetting.remove_override!(setting.name)
-    end
+    SiteSetting.provider.all.each { |setting| SiteSetting.remove_override!(setting.name) }
 
     # very expensive IO operations
     SiteSetting.automatically_download_gravatars = false
@@ -152,41 +145,27 @@ module TestSetup
     # Don't queue badge grant in test mode
     BadgeGranter.disable_queue
 
-    # Make sure the default Post and Topic bookmarkables are registered
-    Bookmark.reset_bookmarkables
-
-    # Make sure only the default category and tag hashtag data sources are registered.
-    HashtagAutocompleteService.clear_data_sources
-
     OmniAuth.config.test_mode = false
   end
 end
 
-TestProf::BeforeAll.configure do |config|
-  config.before(:begin) do
-    TestSetup.test_setup
-  end
-end
+TestProf::BeforeAll.configure { |config| config.before(:begin) { TestSetup.test_setup } }
 
-if ENV['PREFABRICATION'] == '0'
+if ENV["PREFABRICATION"] == "0"
   module Prefabrication
     def fab!(name, &blk)
       let!(name, &blk)
     end
   end
 
-  RSpec.configure do |config|
-    config.extend Prefabrication
-  end
+  RSpec.configure { |config| config.extend Prefabrication }
 else
-  TestProf::LetItBe.configure do |config|
-    config.alias_to :fab!, refind: true
-  end
+  TestProf::LetItBe.configure { |config| config.alias_to :fab!, refind: true }
 end
 
 RSpec.configure do |config|
-  config.fail_fast = ENV['RSPEC_FAIL_FAST'] == "1"
-  config.silence_filter_announcements = ENV['RSPEC_SILENCE_FILTER_ANNOUNCEMENTS'] == "1"
+  config.fail_fast = ENV["RSPEC_FAIL_FAST"] == "1"
+  config.silence_filter_announcements = ENV["RSPEC_SILENCE_FILTER_ANNOUNCEMENTS"] == "1"
   config.extend RedisSnapshotHelper
   config.include Helpers
   config.include MessageBus
@@ -200,13 +179,25 @@ RSpec.configure do |config|
   config.include OneboxHelpers
   config.include FastImageHelpers
   config.mock_framework = :mocha
-  config.order = 'random'
+  config.order = "random"
   config.infer_spec_type_from_file_location!
 
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false
   # instead of true.
   config.use_transactional_fixtures = true
+
+  # Sometimes you may have a large string or object that you are comparing
+  # with some expectation, and you want to see the full diff between actual
+  # and expected without rspec truncating 90% of the diff. Setting the
+  # max_formatted_output_length to nil disables this truncation completely.
+  #
+  # c.f. https://www.rubydoc.info/gems/rspec-expectations/RSpec/Expectations/Configuration#max_formatted_output_length=-instance_method
+  if ENV["RSPEC_DISABLE_DIFF_TRUNCATION"]
+    config.expect_with :rspec do |expectation|
+      expectation.max_formatted_output_length = nil
+    end
+  end
 
   # If true, the base class of anonymous controllers will be inferred
   # automatically. This will be the default behavior in future versions of
@@ -227,7 +218,7 @@ RSpec.configure do |config|
     # Ugly, but needed until we have a user creator
     User.skip_callback(:create, :after, :ensure_in_trust_level_group)
 
-    DiscoursePluginRegistry.reset! if ENV['LOAD_PLUGINS'] != "1"
+    DiscoursePluginRegistry.reset! if ENV["LOAD_PLUGINS"] != "1"
     Discourse.current_user_provider = TestCurrentUserProvider
 
     SiteSetting.refresh!
@@ -244,43 +235,84 @@ RSpec.configure do |config|
 
     SiteSetting.provider = TestLocalProcessProvider.new
 
-    WebMock.disable_net_connect!(
-      allow_localhost: true,
-      allow: [Webdrivers::Chromedriver.base_url]
-    )
+    WebMock.disable_net_connect!(allow_localhost: true, allow: [Webdrivers::Chromedriver.base_url])
+
+    if ENV["CAPBYARA_DEFAULT_MAX_WAIT_TIME"].present?
+      Capybara.default_max_wait_time = ENV["CAPBYARA_DEFAULT_MAX_WAIT_TIME"].to_i
+    end
+
+    Capybara.threadsafe = true
+    Capybara.disable_animation = false
+
+    # Click offsets is calculated from top left of element
+    Capybara.w3c_click_offset = false
 
     Capybara.configure do |capybara_config|
       capybara_config.server_host = "localhost"
-      capybara_config.server_port = 31337
+      capybara_config.server_port = 31_337 + ENV["TEST_ENV_NUMBER"].to_i
     end
 
-    chrome_browser_options = Selenium::WebDriver::Chrome::Options.new(
-      logging_prefs: { "browser" => "ALL", "driver" => "ALL" }
-    ).tap do |options|
-      options.add_argument("--window-size=1400,1400")
-      options.add_argument("--no-sandbox")
-      options.add_argument("--disable-dev-shm-usage")
+    module IgnoreUnicornCapturedErrors
+      def raise_server_error!
+        super
+      rescue EOFError, Errno::ECONNRESET, Errno::EPIPE, Errno::ENOTCONN => e
+        # Ignore these exceptions - caused by client. Handled by unicorn in dev/prod
+        # https://github.com/defunkt/unicorn/blob/d947cb91cf/lib/unicorn/http_server.rb#L570-L573
+      end
     end
+
+    Capybara::Session.class_eval { prepend IgnoreUnicornCapturedErrors }
+
+    # The valid values for SELENIUM_BROWSER_LOG_LEVEL are:
+    #
+    # OFF
+    # SEVERE
+    # WARNING
+    # INFO
+    # DEBUG
+    # ALL
+    browser_log_level = ENV["SELENIUM_BROWSER_LOG_LEVEL"] || "SEVERE"
+    chrome_browser_options =
+      Selenium::WebDriver::Chrome::Options
+        .new(logging_prefs: { "browser" => browser_log_level, "driver" => "ALL" })
+        .tap do |options|
+          options.add_argument("--window-size=1400,1400")
+          options.add_argument("--no-sandbox")
+          options.add_argument("--disable-dev-shm-usage")
+          options.add_argument("--mute-audio")
+        end
 
     Capybara.register_driver :selenium_chrome do |app|
-      Capybara::Selenium::Driver.new(
-        app,
-        browser: :chrome,
-        capabilities: chrome_browser_options,
-      )
+      Capybara::Selenium::Driver.new(app, browser: :chrome, options: chrome_browser_options)
     end
 
     Capybara.register_driver :selenium_chrome_headless do |app|
       chrome_browser_options.add_argument("--headless")
 
-      Capybara::Selenium::Driver.new(
-        app,
-        browser: :chrome,
-        capabilities: chrome_browser_options,
-      )
+      Capybara::Selenium::Driver.new(app, browser: :chrome, options: chrome_browser_options)
     end
 
-    if ENV['ELEVATED_UPLOADS_ID']
+    mobile_chrome_browser_options =
+      Selenium::WebDriver::Chrome::Options
+        .new(logging_prefs: { "browser" => "INFO", "driver" => "ALL" })
+        .tap do |options|
+          options.add_argument("--window-size=390,950")
+          options.add_argument("--no-sandbox")
+          options.add_argument("--disable-dev-shm-usage")
+          options.add_emulation(device_name: "iPhone 12 Pro")
+          options.add_argument("--mute-audio")
+        end
+
+    Capybara.register_driver :selenium_mobile_chrome do |app|
+      Capybara::Selenium::Driver.new(app, browser: :chrome, options: mobile_chrome_browser_options)
+    end
+
+    Capybara.register_driver :selenium_mobile_chrome_headless do |app|
+      mobile_chrome_browser_options.add_argument("--headless")
+      Capybara::Selenium::Driver.new(app, browser: :chrome, options: mobile_chrome_browser_options)
+    end
+
+    if ENV["ELEVATED_UPLOADS_ID"]
       DB.exec "SELECT setval('uploads_id_seq', 10000)"
     else
       DB.exec "SELECT setval('uploads_id_seq', 1)"
@@ -316,9 +348,7 @@ RSpec.configure do |config|
   end
 
   config.after(:suite) do
-    if SpecSecureRandom.value
-      FileUtils.remove_dir(file_from_fixtures_tmp_folder, true)
-    end
+    FileUtils.remove_dir(file_from_fixtures_tmp_folder, true) if SpecSecureRandom.value
   end
 
   config.before :each, &TestSetup.method(:test_setup)
@@ -327,7 +357,8 @@ RSpec.configure do |config|
     before_event_count = DiscourseEvent.events.values.sum(&:count)
     example.run
     after_event_count = DiscourseEvent.events.values.sum(&:count)
-    expect(before_event_count).to eq(after_event_count), "DiscourseEvent registrations were not cleaned up"
+    expect(before_event_count).to eq(after_event_count),
+    "DiscourseEvent registrations were not cleaned up"
   end
 
   config.before :each do
@@ -336,28 +367,31 @@ RSpec.configure do |config|
   end
 
   # Match the request hostname to the value in `database.yml`
-  config.before(:all, type: [:request, :multisite, :system]) { host! "test.localhost" }
-  config.before(:each, type: [:request, :multisite, :system]) { host! "test.localhost" }
+  config.before(:all, type: %i[request multisite system]) { host! "test.localhost" }
+  config.before(:each, type: %i[request multisite system]) { host! "test.localhost" }
 
   last_driven_by = nil
   config.before(:each, type: :system) do |example|
     if example.metadata[:js]
-      driver = "selenium_chrome"
-      driver += "_headless" unless ENV["SELENIUM_HEADLESS"] == "0"
-      driven_by driver.to_sym
+      driver = [:selenium]
+      driver << :mobile if example.metadata[:mobile]
+      driver << :chrome
+      driver << :headless unless ENV["SELENIUM_HEADLESS"] == "0"
+      driven_by driver.join("_").to_sym
     end
     setup_system_test
   end
 
   config.after(:each, type: :system) do |example|
+    lines = RSpec.current_example.metadata[:extra_failure_lines]
+
     # This is disabled by default because it is super verbose,
     # if you really need to dig into how selenium is communicating
     # for system tests then enable it.
     if ENV["SELENIUM_VERBOSE_DRIVER_LOGS"]
-      puts "~~~~~~ DRIVER LOGS: ~~~~~~~"
-      page.driver.browser.logs.get(:driver).each do |log|
-        puts log.message
-      end
+      lines << "~~~~~~~ DRIVER LOGS ~~~~~~~"
+      page.driver.browser.logs.get(:driver).each { |log| lines << log.message }
+      lines << "~~~~~ END DRIVER LOGS ~~~~~"
     end
 
     # Recommended that this is not disabled, since it makes debugging
@@ -365,32 +399,37 @@ RSpec.configure do |config|
     if ENV["SELENIUM_DISABLE_VERBOSE_JS_LOGS"].blank?
       if example.exception
         skip_js_errors = false
+
         if example.exception.kind_of?(RSpec::Core::MultipleExceptionError)
-          puts "~~~~~~ SYSTEM TEST ERRORS: ~~~~~~~"
-          example.exception.all_exceptions.each do |ex|
-            puts ex.message
-          end
+          lines << "~~~~~~~ SYSTEM TEST ERRORS ~~~~~~~"
+          example.exception.all_exceptions.each { |ex| lines << ex.message }
+          lines << "~~~~~ END SYSTEM TEST ERRORS ~~~~~"
 
           skip_js_errors = true
         end
 
         if !skip_js_errors
-          puts "~~~~~~ JS ERRORS: ~~~~~~~"
-          page.driver.browser.logs.get(:browser).each do |log|
-            puts log.message
+          lines << "~~~~~~~ JS LOGS ~~~~~~~"
+          logs = page.driver.browser.logs.get(:browser)
+          if logs.empty?
+            lines << "(no logs)"
+          else
+            logs.each { |log| lines << log.message }
           end
+          lines << "~~~~~ END JS LOGS ~~~~~"
         end
       end
     end
+
+    Discourse.redis.flushdb
   end
 
   config.before(:each, type: :multisite) do
     Rails.configuration.multisite = true # rubocop:disable Discourse/NoDirectMultisiteManipulation
 
-    RailsMultisite::ConnectionManagement.config_filename =
-      "spec/fixtures/multisite/two_dbs.yml"
+    RailsMultisite::ConnectionManagement.config_filename = "spec/fixtures/multisite/two_dbs.yml"
 
-    RailsMultisite::ConnectionManagement.establish_connection(db: 'default')
+    RailsMultisite::ConnectionManagement.establish_connection(db: "default")
   end
 
   config.after(:each, type: :multisite) do
@@ -424,7 +463,6 @@ RSpec.configure do |config|
       end
     end
   end
-
 end
 
 class TrackTimeStub
@@ -442,9 +480,7 @@ def global_setting(name, value)
 
   GlobalSetting.stubs(name).returns(value)
 
-  before_next_spec do
-    GlobalSetting.reset_s3_cache!
-  end
+  before_next_spec { GlobalSetting.reset_s3_cache! }
 end
 
 def set_cdn_url(cdn_url)
@@ -556,22 +592,19 @@ def create_request_env(path: nil)
 end
 
 def create_auth_cookie(token:, user_id: nil, trust_level: nil, issued_at: Time.current)
-  data = {
-    token: token,
-    user_id: user_id,
-    trust_level: trust_level,
-    issued_at: issued_at.to_i
-  }
+  data = { token: token, user_id: user_id, trust_level: trust_level, issued_at: issued_at.to_i }
   jar = ActionDispatch::Cookies::CookieJar.build(ActionDispatch::TestRequest.create, {})
   jar.encrypted[:_t] = { value: data }
   CGI.escape(jar[:_t])
 end
 
 def decrypt_auth_cookie(cookie)
-  ActionDispatch::Cookies::CookieJar
-    .build(ActionDispatch::TestRequest.create, { _t: cookie })
-    .encrypted[:_t]
-    .with_indifferent_access
+  ActionDispatch::Cookies::CookieJar.build(
+    ActionDispatch::TestRequest.create,
+    { _t: cookie },
+  ).encrypted[
+    :_t
+  ].with_indifferent_access
 end
 
 class SpecSecureRandom
